Move `NdlaMiddleware`, `Service` and `Routes` to `network`
More tapir generalization!
