+++ frontpage-api/src/main/scala/no/ndla/frontpageapi/controller/FilmPageController.scala
@@ -13,6 +13,7 @@ import no.ndla.frontpageapi.auth.UserInfo
 import no.ndla.frontpageapi.model.api._
 import no.ndla.frontpageapi.model.domain.Errors.ValidationException
 import no.ndla.frontpageapi.service.{ReadService, WriteService}
+import no.ndla.network.tapir.Service
 import no.ndla.network.tapir.TapirErrors.errorOutputs
 import sttp.tapir._
 import sttp.tapir.generic.auto._
