+++ network/src/main/scala/no/ndla/network/tapir/NdlaMiddleware.scala
@@ -5,7 +5,7 @@
  * See LICENSE
  */
 
-package no.ndla.frontpageapi.controller
+package no.ndla.network.tapir
 
 import cats.effect.IO
 import no.ndla.common.CorrelationID
